+++ blockboard/src/main/java/com/board/project/blockboard/dto/UserDTO.java
@@ -1,47 +1,13 @@
 package com.board.project.blockboard.dto;
 
+import lombok.Data;
+
+@Data
 public class UserDTO {
     private String user_id;
     private int com_id;
     private String user_pwd;
-
-    public String getUser_pwd() {
-        return user_pwd;
-    }
-
-    public void setUser_pwd(String user_pwd) {
-        this.user_pwd = user_pwd;
-    }
-
     private String user_type;
-
-    public UserDTO(String user_id) {
-        this.user_id = user_id;
-    }
-    public String getUser_id() {
-        return user_id;
-    }
-
-    public void setUser_id(String user_id) {
-        this.user_id = user_id;
-    }
-
-    public int getCom_id() {
-        return com_id;
-    }
-
-    public void setCom_id(int com_id) {
-        this.com_id = com_id;
-    }
-
-    public String getUser_type() {
-        return user_type;
-    }
-
-    public void setUser_type(String user_type) {
-        this.user_type = user_type;
-    }
-
 
     @Override
     public String toString() {
