+++ blockboard/src/main/java/com/board/project/blockboard/mapper/UserMapper.java
@@ -9,6 +9,5 @@ import java.util.List;
 @Repository
 @Mapper
 public interface UserMapper {
-    public List<UserDTO> allUser();
-    public UserDTO selectUser(String user_id);
+    public UserDTO getUser(String user_id);
 }
