+++ blockboard/src/main/resources/mybatis/mapper/UserMapper.xml
@@ -3,11 +3,8 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="com.board.project.blockboard.mapper.UserMapper">
-    <select id="allUser" resultType="com.board.project.blockboard.dto.UserDTO">
-        SELECT * FROM Users;
-    </select>
 
-    <select id="selectUser" parameterType="string" resultType="com.board.project.blockboard.dto.UserDTO">
+    <select id="getUser" parameterType="string" resultType="com.board.project.blockboard.dto.UserDTO">
         SELECT *
         FROM Users
         WHERE user_id=#{user_id};
