+++ blockboard/src/main/webapp/WEB-INF/jsp/login.jsp
@@ -2,14 +2,11 @@
            pageEncoding="UTF-8"%>
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
+<script src="http://code.jquery.com/jquery-1.10.2.js"></script>
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=EUC-KR">
-   <script src="http://code.jquery.com/jquery-1.10.2.js"></script>
    <title>로그인 화면</title>
-
-   <!-- css 파일 나중에 넣기 -->
-
 </head>
 
 <body>
