--- blockboard/src/main/webapp/WEB-INF/jsp/post.jsp
@@ -1,16 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<html>
-<head>
-	<meta http-equiv="Content-Type" content="text/html; charset=EUC-KR">
-	<title>게시글</title>
-</head>
-
-<body>
-    <h2>게시글 조회 페이지</h2>
-	<button>댓글 등록</button>
-	<button onclick="location='게시판.jsp'">뒤로 가기</button>
-</body>
-
-</html>
