--- blockboard/src/main/webapp/WEB-INF/jsp/writeform.jsp
@@ -1,21 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<html>
-<head>
-	<meta http-equiv="Content-Type" content="text/html; charset=EUC-KR">
-	<title>글쓰기</title>
-</head>
-
-<body>
-    <script>
-        function loadURL()
-        {
-            var url = document.all.url;
-
-    <h2>글쓰기 페이지</h2>
-	<button>올리기</button>
-	<button onclick="location='게시판.jsp'">뒤로 가기</button>
-</body>
-
-</html>
